core/risk_manager.py
"""
Risk Manager Module - Handles risk assessment and management for the immune system.

This module implements risk calculation, exposure monitoring, and drawdown protection
mechanisms to ensure trading safety.
"""
from typing import Dict, List, Optional
from dataclasses import dataclass
import logging
import numpy as np

logger = logging.getLogger(__name__)

@dataclass
class PositionRisk:
    """Container for position-specific risk metrics."""
    exposure: float
    volatility_risk: float
    correlation_risk: float
    total_risk: float

class RiskManager:
    """
    Handles risk assessment and management for trading positions.
    Implements various risk metrics calculations and protection mechanisms.
    """

    def __init__(self, config: Optional[Dict] = None):
        """
        Initialize the risk manager with configuration parameters.

        Args:
            config: Optional configuration dictionary with risk parameters
        """
        self._config = config or {
            'max_position_size': 0.1,  # 10% of capital
            'max_total_exposure': 0.2,  # 20% of capital
            'correlation_threshold': 0.7,
            'max_drawdown': 0.1,  # 10% drawdown limit
            'volatility_multiplier': 2.0
        }
        logger.info("Risk Manager initialized with config: %s", self._config)

    def calculate_position_risk(
        self,
        position: Dict,
        market_data: Dict,
        portfolio: Dict
    ) -> PositionRisk:
        """
        Calculate comprehensive risk metrics for a single position.

        Args:
            position: Position details including size and entry price
            market_data: Current market data including volatility
            portfolio: Current portfolio state including other positions

        Returns:
            PositionRisk object containing calculated risk metrics
        """
        exposure = self._calculate_exposure(position, portfolio)
        vol_risk = self._calculate_volatility_risk(position, market_data)
        corr_risk = self._calculate_correlation_risk(position, portfolio)
        
        total_risk = self._aggregate_risk_metrics(
            exposure,
            vol_risk,
            corr_risk
        )

        return PositionRisk(
            exposure=exposure,
            volatility_risk=vol_risk,
            correlation_risk=corr_risk,
            total_risk=total_risk
        )

    def check_risk_limits(
        self,
        position_risk: PositionRisk,
        portfolio: Dict
    ) -> bool:
        """
        Check if position risks are within acceptable limits.

        Args:
            position_risk: Calculated position risk metrics
            portfolio: Current portfolio state

        Returns:
            True if risks are acceptable, False otherwise
        """
        if position_risk.total_risk > self._config['max_position_size']:
            logger.warning("Position risk exceeds maximum allowed")
            return False

        total_exposure = self._calculate_total_exposure(portfolio)
        if total_exposure > self._config['max_total_exposure']:
            logger.warning("Total exposure exceeds maximum allowed")
            return False

        return True

    def calculate_drawdown_protection(
        self,
        position: Dict,
        market_data: Dict
    ) -> Dict[str, float]:
        """
        Calculate protective measures for drawdown prevention.

        Args:
            position: Position details
            market_data: Current market data including volatility

        Returns:
            Dictionary containing protection levels (stop loss, etc.)
        """
        volatility = market_data.get('volatility', 0.0)
        current_price = market_data.get('price', position['entry_price'])
        
        # Calculate dynamic protection levels based on volatility
        stop_distance = volatility * self._config['volatility_multiplier']
        
        # Calculate risk level based on volatility and price movement
        risk_level = min(1.0, stop_distance * 2)  # Normalize between 0 and 1
        
        return {
            'stop_loss': current_price * (1 - stop_distance),
            'warning_level': current_price * (1 - stop_distance * 0.7),
            'risk_level': risk_level
        }

    def _calculate_exposure(self, position: Dict, portfolio: Dict) -> float:
        """Calculate position exposure relative to portfolio."""
        position_value = position['size'] * position.get('current_price', 
                                                       position['entry_price'])
        portfolio_value = portfolio.get('total_value', 0)
        return position_value / portfolio_value if portfolio_value else 0

    def _calculate_volatility_risk(
        self,
        position: Dict,
        market_data: Dict
    ) -> float:
        """Calculate risk based on market volatility."""
        volatility = market_data.get('volatility', 0.0)
        return volatility * position['size']

    def _calculate_correlation_risk(
        self,
        position: Dict,
        portfolio: Dict
    ) -> float:
        """Calculate risk based on asset correlation with portfolio."""
        # TODO: Implement correlation calculation
        return 0.0

    def _calculate_total_exposure(self, portfolio: Dict) -> float:
        """Calculate total portfolio exposure."""
        return sum(
            pos['size'] * pos.get('current_price', pos['entry_price'])
            for pos in portfolio.get('positions', [])
        ) / portfolio.get('total_value', 1)

    def _aggregate_risk_metrics(
        self,
        exposure: float,
        volatility_risk: float,
        correlation_risk: float
    ) -> float:
        """
        Aggregate different risk metrics into a single risk score.

        Uses a weighted average of different risk metrics to produce
        a final risk score between 0 and 1.
        """
        weights = {
            'exposure': 0.4,
            'volatility': 0.4,
            'correlation': 0.2
        }
        
        return (
            exposure * weights['exposure'] +
            volatility_risk * weights['volatility'] +
            correlation_risk * weights['correlation']
        )
